Reorganize the code and add service removal
